scam_message_ocr/scam_ocr_detector/requirements.txt
# ---------- core ----------
numpy==1.26.4

# ---------- image / OCR ----------
opencv-python==4.6.0.66
pillow>=10.0.0

paddlepaddle==2.6.2
paddleocr==2.7.3

# ---------- LLM ----------
torch>=2.1.0
transformers>=4.46.0
accelerate>=0.33.0
sentencepiece>=0.2.0
safetensors>=0.4.3

# ---------- HF download ----------
huggingface-hub>=0.24.0
hf_transfer

# ---------- utils ----------
requests>=2.32.0
python-dotenv>=1.0.1
json-repair